+++ data_pipeline/code/url_source_validator.py
@@ -5,8 +5,8 @@ from loguru import logger
 import pandas as pd
 import re
 
-from url_source import UrlSource
-import udatetime
+from .url_source import UrlSource
+from .shared import udatetime
 
 class UrlSourceValidator():
