+++ data_pipeline/code/x_html_compare.py
@@ -4,7 +4,7 @@
 
 
 import re
-from directory_cache import DirectoryCache
+from .directory_cache import DirectoryCache
 
 def clean_content(content: bytes) -> bytes:
