+++ data_pipeline/tests/test_monitor.py
@@ -1,15 +1,11 @@
 #
 # working file for testing git support  .  should be a unit test
 #
-import sys
-import os
 import subprocess
 import time
+import sys
 
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('..'))
-
-from util import monitor_start, monitor_check
+from ..code.shared.util_git import monitor_start, monitor_check
 
 def test_monitor():
