+++ data_pipeline/tests/test_sources.py
@@ -1,24 +1,18 @@
 #
 # working file for testing the sources  .  should be a unit test
 #
-import sys
-import os
-
 from loguru import logger
 
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('..'))
-
-from directory_cache import DirectoryCache
-from url_source import UrlSource
-from url_source_manager import UrlSourceManager
+from ..code.directory_cache import DirectoryCache
+from ..code.url_source import UrlSource
+from ..code.url_source_manager import UrlSourceManager
 
 def test_load():
 
     cache_sources = DirectoryCache("c:\\temp\\sources")
 
     manager = UrlSourceManager(cache_sources)
-    sources = manager.update_sources("test")
+    manager.update_sources("test")
 
     logger.info("if your source was loaded and it didn't print error, it should be fine")
